+++ example/youtube_integration_example.dart
@@ -1,6 +1,0 @@
-import 'package:youtube_integration/youtube_integration.dart';
-
-void main() {
-  var awesome = Awesome();
-  print('awesome: ${awesome.isAwesome}');
-}