+++ lib/models/video_comments/author_channel_id.dart
@@ -1,0 +1,15 @@
+class AuthorChannelId {
+  String? value;
+
+  AuthorChannelId({this.value});
+
+  factory AuthorChannelId.fromJson(Map<String, dynamic> json) {
+    return AuthorChannelId(
+      value: json['value'] as String?,
+    );
+  }
+
+  Map<String, dynamic> toJson() => {
+        'value': value,
+      };
+}