+++ CHANGELOG.md
@@ -1,3 +1,13 @@
-## 1.0.0
+# Changelog
 
-- Initial version.
+All notable changes to this project will be documented in this file.
+
+## [1.0.0] - 2025-04-09
+### Added
+- Initial release of the `youtube_integration` package.
+- Fetch channel videos.
+- Fetch channel playlists.
+- Fetch videos from a playlist.
+- Search for videos.
+- Fetch video details and comments.
+- Fetch YouTube channel details.